+++ webapi/src/test/scala/org/knora/webapi/responders/v2/OntologyResponderV2Spec.scala
@@ -102,18 +102,6 @@ class OntologyResponderV2Spec extends CoreSpec() with ImplicitSender {
   val anythingHasStandoffLinkTo: IRI       = "http://api.knora.org/ontology/knora-api/v2#hasStandoffLinkTo"
   val anythingHasStandoffLinkToValue: IRI  = "http://api.knora.org/ontology/knora-api/v2#hasStandoffLinkToValue"
 
-  private def loadInvalidTestData(rdfDataObjs: List[RdfDataObject]): Unit = {
-    appActor ! ResetRepositoryContent(rdfDataObjs)
-    expectMsg(5 minutes, ResetRepositoryContentACK())
-
-    appActor ! LoadOntologiesRequestV2(
-      requestingUser = KnoraSystemInstances.Users.SystemUser
-    )
-
-    val response: SuccessResponseV2 = expectMsgType[SuccessResponseV2](10.seconds)
-    assert(response.message.contains("An error occurred when loading ontologies"))
-  }
-
   "The ontology responder v2" should {
     "not allow a user to create an ontology if they are not a sysadmin or an admin in the ontology's project" in {
       appActor ! CreateOntologyRequestV2(
@@ -6660,413 +6648,6 @@ class OntologyResponderV2Spec extends CoreSpec() with ImplicitSender {
         assert(newAnythingLastModDate.isAfter(anythingLastModDate))
         anythingLastModDate = newAnythingLastModDate
       }
-    }
-
-    "not load an ontology that has no knora-base:attachedToProject" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/onto-without-project.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a class that's missing a cardinality for a link value property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/missing-link-value-cardinality-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a class that's missing a cardinality for a link property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/missing-link-cardinality-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a class with a cardinality whose subject class constraint is incompatible with the class" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-incompatible-with-scc-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource class without an rdfs:label" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-without-label-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource property without an rdfs:label" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/property-without-label-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource class that is also a standoff class" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/resource-class-is-standoff-class-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource class with a cardinality on an undefined property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-missing-property-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource class with a directly defined cardinality on a non-resource property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-non-resource-prop-cardinality-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource class with a cardinality on knora-base:resourceProperty" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-cardinality-on-kbresprop-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource class with a cardinality on knora-base:hasValue" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-cardinality-on-kbhasvalue-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource class with a base class that has a Knora IRI but isn't a resource class" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/resource-class-with-invalid-base-class-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a standoff class with a cardinality on a resource property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/standoff-class-with-resprop-cardinality-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a standoff class with a base class that's not a standoff class" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/standoff-class-with-invalid-base-class-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a property with a subject class constraint of foaf:Person" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-with-non-knora-scc-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a Knora value property with a subject class constraint of knora-base:TextValue" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-with-value-scc-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a property with a subject class constraint of salsah-gui:Guielement" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-with-guielement-scc-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a property with an object class constraint of foaf:Person" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-with-non-knora-occ-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a property whose object class constraint is incompatible with its base property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-with-incompatible-occ-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a class with cardinalities for a link property and a matching link value property, except that the link property isn't really a link property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-misdefined-link-property-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a class with cardinalities for a link property and a matching link value property, except that the link value property isn't really a link value property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-misdefined-link-value-property-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource property with no rdfs:label" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/resource-prop-without-label-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a property that's a subproperty of both knora-base:hasValue and knora-base:hasLinkTo" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-both-value-and-link-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a property that's a subproperty of knora-base:hasFileValue" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/filevalue-prop-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a resource property with a base property that has a Knora IRI but isn't a resource property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/resource-prop-wrong-base-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a cardinality that contains salsah-gui:guiElement" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/cardinality-with-guielement-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load an ontology containing a cardinality that contains salsah-gui:guiAttribute" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/cardinality-with-guiattribute-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing an owl:TransitiveProperty" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/transitive-prop.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology with a class that has cardinalities both on property P and on a subproperty of P" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-inherits-prop-and-subprop-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing mismatched cardinalities for a link property and a link value property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-mismatched-link-cardinalities-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing an invalid cardinality on a boolean property" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/invalid-card-on-boolean-prop.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing a class with a cardinality on a property from a non-shared ontology in another project" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-non-shared-cardinality.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing a class with a base class defined in a non-shared ontology in another project" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/class-with-non-shared-base-class.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing a property with a base property defined in a non-shared ontology in another project" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-with-non-shared-base-prop.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing a property whose subject class constraint is defined in a non-shared ontology in another project" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-with-non-shared-scc.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing a property whose object class constraint is defined in a non-shared ontology in another project" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/prop-with-non-shared-occ.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
-    }
-
-    "not load a project-specific ontology containing a class with two cardinalities that override the same base class cardinality of 1 or 0-1" in {
-      val invalidOnto = List(
-        RdfDataObject(
-          path = "test_data/responders.v2.OntologyResponderV2Spec/conflicting-cardinalities-onto.ttl",
-          name = "http://www.knora.org/ontology/invalid"
-        )
-      )
-
-      loadInvalidTestData(invalidOnto)
     }
 
   }
